+++ db/migrate/20170516092845_add_sitemap_to_page.rb
@@ -1,0 +1,6 @@
+class AddSitemapToPage < ActiveRecord::Migration[5.0]
+  def change
+    add_column :pages, :sitemap, :boolean, after: 'content_disabled'
+    add_index :pages, :sitemap
+  end
+end